feat: first commit

+++ src/main/java/com/projectspring/aula/domain/entity/Cliente.java
@@ -9,14 +9,14 @@ public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private long id;
+    private Long id;
 
     @Column(name = "nome", length = 100)
     private String nome;
 
     public Cliente() {}
 
-    public Cliente(Integer id, String nome) {
+    public Cliente(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
@@ -25,11 +25,11 @@ public class Cliente {
         this.nome = nome;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
